parameters.py moved into hgc_tpg/resolution and utilities moved into hgc_tpg
