+++ hgc_tpg/utilities/functions.py
